Moved a few test things around
