+++ tests/Shared/StandardEnum.cs
@@ -1,0 +1,8 @@
+namespace SurrealDB.Common;
+
+public enum StandardEnum {
+    Zero = 0,
+    One = 1,
+    TwoHundred = 200,
+    NegTwoHundred = -200,
+}
